refactor(types): split 1805-line bag-of-types into cohesive submodules
`types.rs` had become a dumping ground for 15+ unrelated types spanning
buffer groups, search state, buffer metadata, LSP URIs, hover targets,
context menus, theme inspection, drag state, mouse state, layout caches,
and overlay previews.

Split into a `types/` module directory with 11 focused files:
- `buffer_group.rs`  — BufferGroupId, GroupLayoutNode, BufferGroup
- `buffer_meta.rs`   — BufferKind, BufferMetadata
- `context_menu.rs`  — TabContextMenu, FileExplorerContextMenu and items
- `drag.rs`          — TabDropZone, TabDragState
- `hover.rs`         — HoverTarget
- `layout.rs`        — ViewLineMapping, ChromeLayout, WindowLayoutCache, OverlayPreviewState
- `lsp_state.rs`     — LspProgressInfo, LspMenuItem, LspMessageEntry
- `lsp_uri.rs`       — LspUri, file_path_to_lsp_uri* helpers + tests
- `mouse.rs`         — MouseState
- `search_state.rs`  — SearchState, InteractiveReplaceState, EventLineInfo
- `theme.rs`         — CellThemeInfo, ThemeKeyInfo, ThemeInfoPopup

All re-exports preserved in `mod.rs` with original visibility, so no
call sites change.

Also refactors BufferMetadata::is_library_path from 65+ lines of
repeated `if path_str.contains("X") || path_str.contains("Y") { return
true; }` (Copy-Paste Specialist) to two const slices + `any()`.
